requirements.txt
slack_sdk
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
